+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## [1.1.0] - 2026-02-08
+* FEA: Add send new email modal for super admins.
+* FEA: Add email resend functionality.
+* BUG: Remove package-json.lock files.
+
 ## [1.0.0] - 2026-02-08
 * Initial Release.